Filename extension by printer only
Deprecates the PathProviders util class and replaces it with the new PathProviderBuilder.
PathProviderBuilder can be directly used in ApprovalBuilder, in which case the filenameExtension is set just before approval.
Ultimately this makes sure that the filenameExtension is either explicitly defined or is defined by the used Printer.
This is also true if the configured defaultPrinter is used.