Change error message for unapproved case
In case there was no approved value (or that value is empty), the error
message now no longer speaks of an approval mismatch, but of a missing
approval for the received value.
This way the message is more concise, and it is more obvious if the test
case/class got renamed, but not the approval file.