Add new PrintFormat
The new PrintFormat replaces Printer as the main interface to control
printing in ApproveJ. The Printer still exists, but will be a mere
Function in 1.0.

This allows for a clear distinction in the ApproveBuilder's API:
printedAs(PrintFormat) vs printedBy(Printer)/printedBy(Function).
The first overrides the filename extension, the latter ones do not.

This is a breaking change since the Configuration no longer accepts
the defaultPrinter parameter, but only the new defaultPrintFormat!